[tuner]  create interface for constraint generation
